BUG: Fix how `find_hits` matches hits (#120)
so we don't get a precision greater than 1.0

* BUG: Fix matching in find_hits, fixes #119

This fixes how we match hits inside
`metrics.segmentation.ir.find_hits`.

After finding all hits within the window specified
by ``tolerance``, we sort those by difference in time between
the reference boundary and hypothesis boundary.
We then add hits to the output,
but only if we have not already added either the reference boundary
or the hypothesized boundary.
For more detail, see discussion on
https://github.com/vocalpy/vocalpy/issues/119.

* Add test case in test_segmentation/test_ir.py