BUG: Fix cepstrum in spectral.sat, fixes #122
Applies `np.abs` to FFT output before computing log,
so that we replicate what the original SAT code does
when calling the Matlab function `rceps`.

Also adds an `eps` to any zero values in the FFT,
to avoid divide-by-zero errors when taking the log.

Also moves the line computing frequencies `f` down to just above
where it is first used in the function.